Remove all dependencies on Qt from the API module (#102)
Before, many files in the `api` directory imported `app.LmTools` which
has a hard dependency on Qt.  This meant that it was impossible to use
_only_ the API without having Qt installed.

Now, all of the functions that used to be in `app.LmTools` that had no
dependency on Qt are in `util.LmUtils`.  This allows the API to be
used with no Qt dependency.

All files that depend on the moved functions have been updated.